+++ src/proj/EventStore/IStoreEvents.cs
@@ -20,7 +20,7 @@ namespace EventStore
 		/// <param name="id">The value which uniquely identifies the aggregate of the events to be loaded.</param>
 		/// <param name="startingVersion">The version at which the loaded event stream should begin.</param>
 		/// <returns>A stream of committed events for the aggregate indicated.</returns>
-		CommittedEventStream ReadFrom(Guid id, int startingVersion);
+		CommittedEventStream ReadFrom(Guid id, long startingVersion);
 
 		/// <summary>
 		/// Writes the stream of uncommitted events to persistent storage.
